Use Import::BaseAction as base-class
... for Import::CollectAction and Import::CreateAction
